Prime Number Algorithm
There is no specific rule of prime numbers. Therefore, it is very
difficult to understand that it is a prime number. This algorithm status
can be queried through any number. If you enter a prime number tells
you. If the entered number is not prime, the next prime number tells
you. Easy and convenient!